10k River Gorge Thoughts + changes
Commented out the javascript_dir in the config.rb file in the theme.
Created a div.article__summary to insert pictures with on my excerpt.
This is also made to center the picture.  Anything put in the summary
for the excerpt is going to be put onto the main page at the top.
Created the {{pagination}} arrows.  This makes it to where we can
navigate pages when there are more articles to read.